refactor: extract SummaryFormatter to _formatter.py
