+++ test/test_formatter.py
@@ -1,7 +1,8 @@
 import os
 from unittest import mock
 import pytest
-from con_duct.duct_main import Report, SummaryFormatter
+from con_duct._formatter import SummaryFormatter
+from con_duct.duct_main import Report
 
 GREEN_START = SummaryFormatter.COLOR_SEQ % SummaryFormatter.GREEN
 RED_START = SummaryFormatter.COLOR_SEQ % SummaryFormatter.RED
